// src/types/player.ts
export interface PlayerData {
  id: string;
  name: string;
  rank: number;
  rating: string;
  team: string;
  pts: number;
  reb: number;
  ast: number;
  stl: number;
  blk: number;
  gradient: string;
  color: string;
  position: string;
  age: number;
  gp: number;
  min: number;
  fgPct: number;
  threePct: number;
  plusMinus: number;
}